Fix next on netlify deployment issues with i18n.
