+++ content/imprint.json
@@ -1,29 +1,29 @@
 {
-  "de": {
-    "title": "Impressum",
-    "organizationName": "Jakob Kolb",
-    "organizationAddress": "Fuldastraße 19\n12045 Berlin",
-    "email": "lieber-zusammen@posteo.de",
-    "backgroundImage": "/img/landing.jpg"
-  },
   "en": {
     "title": "Imprint",
     "organizationName": "Jakob Kolb",
-    "organizationAddress": "Fuldastraße 19\n12045 Berlin",
+    "organizationAddress": "Fuldastraße 19\n\n12045 Berlin",
+    "email": "lieber-zusammen@posteo.de",
+    "backgroundImage": "/img/landing.jpg"
+  },
+  "de": {
+    "title": "Impressum",
+    "organizationName": "Jakob Kolb",
+    "organizationAddress": "Fuldastraße 19\n\n12045 Berlin",
     "email": "lieber-zusammen@posteo.de",
     "backgroundImage": "/img/landing.jpg"
   },
   "de-simple": {
     "title": "Impressum",
     "organizationName": "Jakob Kolb",
-    "organizationAddress": "Fuldastraße 19\n12045 Berlin",
+    "organizationAddress": "Fuldastraße 19\n\n12045 Berlin",
     "email": "lieber-zusammen@posteo.de",
     "backgroundImage": "/img/landing.jpg"
   },
   "en-simple": {
     "title": "Imprint",
     "organizationName": "Jakob Kolb",
-    "organizationAddress": "Fuldastraße 19\n12045 Berlin",
+    "organizationAddress": "Fuldastraße 19\n\n12045 Berlin",
     "email": "lieber-zusammen@posteo.de",
     "backgroundImage": "/img/landing.jpg"
   }
