+++ netlify.toml
@@ -1,0 +1,4 @@
+[build]
+  command   = "yarn export"
+  functions = "out_functions"
+  publish   = "out_publish"
