+++ package.json
@@ -7,8 +7,9 @@
     "test": "yarn jest",
     "dev": "next",
     "build": "next build",
+    "postbuild": "yarn next-on-netlify",
     "start": "next start",
-    "export": "yarn build",
+    "export": "yarn build && yarn postbuild",
     "install:clean": "rm -rf node_modules/ && rm -rf package-lock.json && rm -rf .next/ && npm install && npm run dev"
   },
   "author": "Jakob J. Kolb",
@@ -29,6 +30,7 @@
     "next": "^10.0.3",
     "next-compose-plugins": "2.2.0",
     "next-images": "1.3.1",
+    "next-on-netlify": "github:netlify/next-on-netlify#ll/i18n",
     "nextjs-redirect": "2.0.2",
     "node-fetch": "^2.6.1",
     "node-sass": "4.14.1",
